move system settings structs to another file
