+++ crates/gpui/src/platform/windows.rs
@@ -3,6 +3,7 @@ mod dispatcher;
 mod display;
 mod events;
 mod platform;
+mod system_settings;
 mod util;
 mod window;
 
@@ -11,6 +12,7 @@ pub(crate) use dispatcher::*;
 pub(crate) use display::*;
 pub(crate) use events::*;
 pub(crate) use platform::*;
+pub(crate) use system_settings::*;
 pub(crate) use util::*;
 pub(crate) use window::*;
