+++ crates/gpui/src/platform/windows/platform.rs
@@ -3,7 +3,7 @@
 
 use std::{
     cell::{Cell, RefCell},
-    ffi::{c_uint, c_void, OsString},
+    ffi::{c_void, OsString},
     mem::transmute,
     os::windows::ffi::{OsStrExt, OsStringExt},
     path::{Path, PathBuf},
@@ -49,21 +49,6 @@ pub(crate) struct WindowsPlatform {
     dispatch_event: OwnedHandle,
 }
 
-/// Windows settings pulled from SystemParametersInfo
-/// https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-systemparametersinfow
-#[derive(Default, Debug)]
-pub(crate) struct WindowsPlatformSystemSettings {
-    pub(crate) mouse_wheel_settings: MouseWheelSettings,
-}
-
-#[derive(Default, Debug, Clone, Copy)]
-pub(crate) struct MouseWheelSettings {
-    /// SEE: SPI_GETWHEELSCROLLCHARS
-    pub(crate) wheel_scroll_chars: u32,
-    /// SEE: SPI_GETWHEELSCROLLLINES
-    pub(crate) wheel_scroll_lines: u32,
-}
-
 pub(crate) struct WindowsPlatformState {
     callbacks: PlatformCallbacks,
     pub(crate) settings: WindowsPlatformSystemSettings,
@@ -79,65 +64,6 @@ struct PlatformCallbacks {
     app_menu_action: Option<Box<dyn FnMut(&dyn Action)>>,
     will_open_app_menu: Option<Box<dyn FnMut()>>,
     validate_app_menu_command: Option<Box<dyn FnMut(&dyn Action) -> bool>>,
-}
-
-impl WindowsPlatformSystemSettings {
-    fn new() -> Self {
-        let mut settings = Self::default();
-        settings.init();
-        settings
-    }
-
-    fn init(&mut self) {
-        self.mouse_wheel_settings.update();
-    }
-}
-
-impl MouseWheelSettings {
-    pub fn update(&mut self) -> (Option<u32>, Option<u32>) {
-        (
-            self.update_wheel_scroll_chars(),
-            self.update_wheel_scroll_lines(),
-        )
-    }
-
-    fn update_wheel_scroll_chars(&mut self) -> Option<u32> {
-        let mut value = c_uint::default();
-        let result = unsafe {
-            SystemParametersInfoW(
-                SPI_GETWHEELSCROLLCHARS,
-                0,
-                Some((&mut value) as *mut c_uint as *mut c_void),
-                SYSTEM_PARAMETERS_INFO_UPDATE_FLAGS::default(),
-            )
-        };
-
-        if result.log_err() != None && self.wheel_scroll_chars != value {
-            self.wheel_scroll_chars = value;
-            Some(value)
-        } else {
-            None
-        }
-    }
-
-    fn update_wheel_scroll_lines(&mut self) -> Option<u32> {
-        let mut value = c_uint::default();
-        let result = unsafe {
-            SystemParametersInfoW(
-                SPI_GETWHEELSCROLLLINES,
-                0,
-                Some((&mut value) as *mut c_uint as *mut c_void),
-                SYSTEM_PARAMETERS_INFO_UPDATE_FLAGS::default(),
-            )
-        };
-
-        if result.log_err() != None && self.wheel_scroll_lines != value {
-            self.wheel_scroll_lines = value;
-            Some(value)
-        } else {
-            None
-        }
-    }
 }
 
 impl WindowsPlatform {
